Learning day button displayed only for self learning days + disabled based on limits
